+++ src/util/ajax.ts
@@ -91,6 +91,8 @@ export const getReferrer = isWorker() ?
     () => (self as any).worker && (self as any).worker.referrer :
     () => (window.location.protocol === 'blob:' ? window.parent : window).location.href;
 
+export const getProtocolAction = url => config.REGISTERED_PROTOCOLS[url.substring(0, url.indexOf('://'))];
+
 // Determines whether a URL is a file:// URL. This is obviously the case if it begins
 // with file://. Relative URLs are also file:// URLs iff the original document was loaded
 // via a file:// URL.
@@ -223,8 +225,7 @@ export const makeRequest = function(requestParameters: RequestParameters, callba
             return (self as any).worker.actor.send('getResource', requestParameters, callback);
         }
         if (!isWorker()) {
-            const protocol = requestParameters.url.substring(0, requestParameters.url.indexOf('://'));
-            const action = config.REGISTERED_PROTOCOLS[protocol] || makeFetchRequest;
+            const action = getProtocolAction(requestParameters.url) || makeFetchRequest;
             return action(requestParameters, callback);
         }
     }
@@ -255,12 +256,12 @@ export const postData = function(requestParameters: RequestParameters, callback:
     return makeRequest(extend(requestParameters, {method: 'POST'}), callback);
 };
 
-function sameOrigin(url): boolean {
-
-    // string comparison seems a bit of old fashion and awkward, but it is the fastest compared to URL class and regex
+export function sameOrigin(url) {
+ // string comparison seems a bit of old fashion and awkward, but it is the fastest compared to URL class and regex
     const locationObj = window.document.location;
     const expected = `${locationObj.protocol}//${locationObj.host}`;
     return url.toLowerCase().indexOf(expected) === 0;
+
 }
 
 export type ExpiryData = {cacheControl?: string | null; expires?: Date | string | null};