+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ### ✨ Features and improvements
 - [Breaking] Improve control performance by restricting worker count to a max of 1 except safari browser. ([#2354](https://github.com/maplibre/maplibre-gl-js/pull/2354))
+- Improve performance by using HTMLImageElement to download raster source images when refreshExpiredTiles tiles is false ([#2126](https://github.com/maplibre/maplibre-gl-js/pull/2126))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
 
+- Fix `LngLatBounds.extend()` to correctly handle `{ lng: number, lat: number }` coordinates. ([#2425](https://github.com/maplibre/maplibre-gl-js/pull/2425))
 - _...Add new stuff here..._
 
 ## 3.0.0-pre.5